requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
alembic=1.5.8=pypi_0
anyio=2.2.0=py38hecd8cb5_1
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=20.1.0=py38h9ed2024_1
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3eb1b0_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=3.3.0=pyhd3eb1b0_0
brotli=1.0.9=pypi_0
brotlipy=0.7.0=py38h9ed2024_1003
ca-certificates=2021.4.13=hecd8cb5_1
certifi=2020.12.5=py38hecd8cb5_0
cffi=1.14.5=py38h2125817_0
chardet=4.0.0=py38hecd8cb5_1003
click=7.1.2=pypi_0
cryptography=3.4.7=py38h2fd3fbb_0
cycler=0.10.0=pypi_0
dash=1.13.3=pypi_0
dash-core-components=1.10.1=pypi_0
dash-html-components=1.0.3=pypi_0
dash-renderer=1.5.0=pypi_0
dash-table=4.8.1=pypi_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dominate=2.6.0=pypi_0
entrypoints=0.3=py38_0
flasgger=0.9.5=pypi_0
flask=1.1.2=pypi_0
flask-bootstrap4=4.0.2=pypi_0
flask-compress=1.9.0=pypi_0
flask-migrate=2.7.0=pypi_0
flask-sqlalchemy=2.5.1=pypi_0
flask-wtf=0.14.3=pypi_0
future=0.18.2=pypi_0
greenlet=1.0.0=pypi_0
gunicorn=20.1.0=pypi_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py38hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
itsdangerous=1.1.0=pypi_0
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pypi_0
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter-packaging=0.7.12=pyhd3eb1b0_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38hecd8cb5_0
jupyter_server=1.4.1=py38hecd8cb5_0
jupyterlab=3.0.11=pyhd3eb1b0_1
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.4.0=pyhd3eb1b0_0
kiwisolver=1.3.1=pypi_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libsodium=1.0.18=h1de35cc_0
mako=1.1.4=pypi_0
markupsafe=1.1.1=py38h1de35cc_1
matplotlib=3.4.1=pypi_0
mistune=0.8.4=py38h1de35cc_1001
nbclassic=0.2.6=pyhd3eb1b0_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.3.0=py38hecd8cb5_0
numpy=1.19.5=pypi_0
openssl=1.1.1k=h9ed2024_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.3=pypi_0
pandoc=2.12=hecd8cb5_0
pandocfilters=1.4.3=py38hecd8cb5_1
parso=0.8.2=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickle-mixin=1.0.2=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=pypi_0
pip=21.0.1=py38hecd8cb5_0
plotly=4.14.3=pypi_0
plotly-orca=1.2.1=1
ply=3.11=pypi_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
psutil=5.8.0=py38h9ed2024_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.17.3=py38haf1e3a3_0
pysocks=1.7.1=py38_1
python=3.8.8=h88f2d9e_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python-dotenv=0.17.0=pypi_0
python-editor=1.0.4=pypi_0
python-kaleido=0.0.3=py38_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py38h23ab428_1
readline=8.1=h9ed2024_0
requests=2.25.1=pyhd3eb1b0_0
retrying=1.3.3=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.2=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38hecd8cb5_0
six=1.15.0=py38hecd8cb5_0
sklearn=0.0=pypi_0
sniffio=1.2.0=py38hecd8cb5_1
sqlalchemy=1.4.9=pypi_0
sqlite=3.35.4=hce871da_0
terminado=0.9.4=py38hecd8cb5_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py38h9ed2024_0
traitlets=5.0.5=pyhd3eb1b0_0
urllib3=1.26.4=pyhd3eb1b0_0
visitor=0.1.3=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wtforms=2.3.3=pypi_0
xz=5.2.5=h1de35cc_0
yellowbrick=1.3.post1=pypi_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
